requirements.txt
# Core dependencies
pyserial-asyncio>=0.6
click>=8.0
python-dotenv>=0.19
aiohttp>=3.8

# Install the library in development mode
-e ./lib